Development

+++ apps/triplo/src/app/pages/project/project-detail/project-detail.component.ts
@@ -1,26 +1,35 @@
-import {Component, OnInit} from '@angular/core';
+import {Component, Inject, Injector, OnInit} from '@angular/core';
 import {ProjectHttpService} from "../../../services/projects/project-http.service";
-import {CommentInterface, ProjectInterface} from "@triplo/models";
+import {CommentInterface, ProjectInterface, TaskInterface} from "@triplo/models";
 import {ActivatedRoute, Router} from "@angular/router";
 import {CommentHttpService} from "../../../services/comments/comment-http.service";
-import {Observable} from "rxjs";
+import {PolymorpheusComponent} from '@tinkoff/ng-polymorpheus';
+import {Observable, switchMap} from "rxjs";
+import {TaskHttpService} from "../../../services/task/task-http.service";
+import {TuiAlertService, TuiNotification} from "@taiga-ui/core";
+import {ConfirmAlertComponent} from "../../../shared/alert/confirm/confirm-alert.component";
 
 @Component({
   selector: 'triplo-project-detail',
   templateUrl: './project-detail.component.html',
-  styles: [],
 })
 export class ProjectDetailComponent implements OnInit {
   project$!: Observable<ProjectInterface>
   recentComments: CommentInterface[] = []
   comments$!: Observable<CommentInterface[]>
   id!: string
+  $tasks: Observable<TaskInterface[]>;
+  notification: Observable<boolean>
 
   constructor(
+    @Inject(TuiAlertService)
+    private readonly alertService: TuiAlertService,
+    @Inject(Injector) private readonly injector: Injector,
     private route: ActivatedRoute,
     private router: Router,
     private projectService: ProjectHttpService,
-    private commentService: CommentHttpService
+    private commentService: CommentHttpService,
+    private taskService: TaskHttpService
   ) {
   }
 
@@ -32,14 +41,30 @@ export class ProjectDetailComponent implements OnInit {
 
     this.project$ = this.projectService.findProjectById(this.id)
     this.comments$ = this.commentService.getTopLevelComments(this.id)
+    this.$tasks = this.taskService.getTopLevelTasks(this.id);
+
+
+    this.notification = this.alertService.open<boolean>(
+      new PolymorpheusComponent(ConfirmAlertComponent, this.injector),
+      {
+        label: `Are you sure you would like to delete this project?`,
+        status: TuiNotification.Error,
+        autoClose: false,
+      },
+    )
   }
 
   deleteProject() {
-    this.projectService.deleteProject(this.id).subscribe(
-      p => {
-        this.router.navigate(["/Projects"])
+    this.notification.subscribe(b => {
+      if (b) {
+        this.projectService.deleteProject(this.id).subscribe(
+          p => {
+            this.alertService.open('Deleted project', {label: "Success!"}).subscribe()
+            this.router.navigate(["/Projects"])
+          }
+        )
       }
-    )
+    })
   }
 
   back() {
@@ -47,12 +72,19 @@ export class ProjectDetailComponent implements OnInit {
   }
 
   createComment($event: CommentInterface) {
-    // TODO: implement assign proper user
     $event.username = "Monke"
     $event.owner = "638b2dd312a4cfd63a04ba40"
     this.commentService.createComment(this.id, $event).subscribe(data => {
         this.recentComments.push(data)
       }
     );
+  }
+
+  dislike(_id: string) {
+
+  }
+
+  like(_id: string) {
+
   }
 }
